further optimized CreatePathWithQueryString, path never has a querystring portion now that sniff internally uses request parameters too
